Contact validations implemented.

+++ Finance.UI/Controllers/ContactController.cs
@@ -1,7 +1,10 @@
 ﻿using Finance.Core.DTOs;
+using Finance.Core.Validators;
 using Finance.Service;
 using Finance.UI.Views;
+using FluentValidation;
 using System;
+using System.Diagnostics;
 using System.Windows.Forms;
 
 namespace Finance.UI.Controllers
@@ -11,6 +14,8 @@ namespace Finance.UI.Controllers
         private readonly ContactService contService;
         private readonly UserService userService;
         private readonly UserDto loggedUser;
+        private readonly CreateContactValidator createContVal;
+        private readonly UpdateContactValidator updateContVal;
         private ContactView view;
 
         public ContactDto SelectedContDto { get; private set; }
@@ -22,6 +27,8 @@ namespace Finance.UI.Controllers
             contService = new ContactService();
             userService = new UserService();
             loggedUser = userService.GetUser(Environment.UserName);
+            createContVal = new CreateContactValidator();
+            updateContVal = new UpdateContactValidator();
         }
 
         public void Init()
@@ -71,13 +78,21 @@ namespace Finance.UI.Controllers
                     PhoneNumber = view.PhoneNumber.Text
                 };
 
+                createContVal.ValidateAndThrow(createContDto);
+
                 contService.AddCont(createContDto);
                 GetConts();
                 ClearForm();
                 view.ShowMessage("Contact add success.");
             }
-            catch (Exception)
+            catch (ValidationException ex)
             {
+                Debug.WriteLine(ex);
+                view.ShowMessage("Invalid user input.");
+            }
+            catch (Exception ex)
+            {
+                Debug.WriteLine(ex);
                 view.ShowMessage("Contact add failed.");
             }
         }
@@ -86,6 +101,11 @@ namespace Finance.UI.Controllers
         {
             try
             {
+                if (SelectedContDto == null)
+                {
+                    throw new InvalidOperationException();
+                }
+
                 var updateContDto = new UpdateContactDto
                 {
                     ContactId = SelectedContDto.ContactId,
@@ -95,13 +115,26 @@ namespace Finance.UI.Controllers
                     PhoneNumber = view.PhoneNumber.Text
                 };
 
+                updateContVal.ValidateAndThrow(updateContDto);
+
                 contService.UpdateContact(updateContDto);
                 GetConts();
                 ClearForm();
                 view.ShowMessage("Contact update success.");
             }
-            catch (Exception)
+            catch (InvalidOperationException ex)
             {
+                Debug.WriteLine(ex);
+                view.ShowMessage("Please select a contact");
+            }
+            catch (ValidationException ex)
+            {
+                Debug.WriteLine(ex);
+                view.ShowMessage("Invalid user input.");
+            }
+            catch (Exception ex)
+            {
+                Debug.WriteLine(ex);
                 view.ShowMessage("Contact update failed.");
             }
         }
@@ -110,14 +143,25 @@ namespace Finance.UI.Controllers
         {
             try
             {
+                if (SelectedContDto == null)
+                {
+                    throw new InvalidOperationException();
+                }
+
                 var contId = SelectedContDto.ContactId;
                 contService.DeleteCont(contId);
                 GetConts();
                 ClearForm();
                 view.ShowMessage("Contact delete success.");
             }
-            catch (Exception)
+            catch (InvalidOperationException ex)
             {
+                Debug.WriteLine(ex);
+                view.ShowMessage("Please select a contact");
+            }
+            catch (Exception ex)
+            {
+                Debug.WriteLine(ex);
                 view.ShowMessage("Contact delete failed.");
             }
         }
